requirements.txt
# MAE on DTD — reproducible dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.2.0
scipy>=1.10.0
scikit-image>=0.20.0
matplotlib>=3.7.0
Pillow>=9.5.0

# optional: tensorboard
tensorboard>=2.12.0
